# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MinuteOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List compliance alerts
      parameters:
        - name: severity
          in: query
          schema:
            type: ["string", "null"]
        - name: resolved
          in: query
          schema:
            type: ["string", "null"]
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAlertBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        severity:
          type: string
        studentId:
          type: ["integer", "null"]
        studentName:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        serviceRequirementId:
          type: ["integer", "null"]
        message:
          type: string
        suggestedAction:
          type: ["string", "null"]
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        resolvedNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, severity, message, resolved, createdAt]
    ResolveAlertBody:
      type: object
      properties:
        resolvedNote:
          type: ["string", "null"]
      required: []
...